Move filtering function to DarCollectionUtils
